images can be added to products

+++ src/modules/products/controllers/deleteProduct.js
@@ -3,9 +3,14 @@ const { ProductRepository } = require('../product.repository');
 const deleteProduct = async (ctx) => {
   const { productId } = ctx.params;
 
-  const deleted = await ProductRepository.delete(productId);
+  const product = await ProductRepository.findByUUID(productId);
 
-  return deleted ? 'product was deleted' : 'product doesn`t exist';
+  if (product.is_active) {
+    const deleted = await ProductRepository.delete(productId);
+    return deleted;
+  }
+
+  return { result: 'product doesn`t exist' };
 };
 
 module.exports = { deleteProduct };
